Build byte-swapped longs in unsigned arithmetic (LongSwap signed-shift UB)

+++ code/game/q_shared.c
@@ -159,7 +159,8 @@ int    LongSwap (int l)
 	b3 = (l>>16)&255;
 	b4 = (l>>24)&255;
 
-	return ((int)b1<<24) + ((int)b2<<16) + ((int)b3<<8) + b4;
+	// build in unsigned: a signed b1<<24 overflows int for b1 >= 0x80
+	return (int)(((unsigned int)b1<<24) + ((unsigned int)b2<<16) + ((unsigned int)b3<<8) + b4);
 }
 
 int	LongNoSwap (int l)
